requirements.txt
sqlmodel~=0.0.16
fastapi~=0.110.1
SQLAlchemy~=2.0.29
python-dotenv~=1.0.1
alembic~=1.13.1
pyserial~=3.5
pydantic~=2.6.4
pytest~=8.1.1
pydantic-settings
pytest-asyncio
uvicorn[standard]
